updated files

+++ Backend/demo/src/main/java/com/example/Models/CustomerMaster.java
@@ -14,7 +14,7 @@ public class CustomerMaster {
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "customer_id")
-    private int customerId;
+    private Long customerId;
 
     private String firstName;
 
@@ -59,11 +59,11 @@ public class CustomerMaster {
 
     private LocalDate dateOfBirth;
 
-	public int getCustomerId() {
+	public Long getCustomerId() {
 		return customerId;
 	}
 
-	public void setCustomerId(int customerId) {
+	public void setCustomerId(Long customerId) {
 		this.customerId = customerId;
 	}
 
@@ -238,7 +238,7 @@ public class CustomerMaster {
 	public CustomerMaster() {
 		
 	}
-	public CustomerMaster(int customerId, String firstName, String lastName, String addressLine1, String addressLine2,
+	public CustomerMaster(Long customerId, String firstName, String lastName, String addressLine1, String addressLine2,
 			String email, String city, String pincode, String phoneNumber, String mobileNumber, String creditCardType,
 			String creditCardNumber, String drivingLicenseNumber, String idpNumber, String issuedByDL,
 			LocalDate validThroughDL, String passportNumber, LocalDate passportValidThrough, String passportIssuedBy,
@@ -281,5 +281,6 @@ public class CustomerMaster {
 				+ ", dateOfBirth=" + dateOfBirth + "]";
 	}
 
+	
     
 }
